tune-llms/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
trl>=0.7.0
wandb>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
huggingface-hub>=0.17.0
sentencepiece>=0.1.99
protobuf>=4.23.0
PyYAML>=6.0
flask>=2.3.0
flask-cors>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0 